Refactor: Modularize code and add dynamic domain extraction
- Break out monolithic amazon.ts into separate modules:
  - products.ts: Product search and details functions
  - cart.ts: Cart management functions (add, clear, get content)
  - orders.ts: Order history functions

- Add dynamic Amazon domain extraction from cookies:
  - Automatically detects regional Amazon sites (amazon.co.uk, amazon.de, etc.)
  - Falls back to amazon.com if no cookies found
  - All URLs now use the detected domain instead of hardcoded values

- Update all imports and test files to use new module structure
- Add CLAUDE.md documentation for future Claude Code instances
- Fix TypeScript diagnostic issues

This refactoring improves code maintainability and enables the MCP server
to work with any regional Amazon site based on the provided cookies.
