+++ src/config.ts
@@ -2,7 +2,7 @@ import { loadAmazonCookiesFile } from './utils.js'
 
 const __dirname = new URL('.', import.meta.url).pathname
 
-export const IS_BROWSER_VISIBLE = true
+export const IS_BROWSER_VISIBLE = false
 
 /** Use local mock files instead of live scraping */
 export const USE_MOCKS = false
@@ -31,3 +31,50 @@ export const AMAZON_COOKIES: {
   storeId: string | null
   value: string
 }[] = loadAmazonCookiesFile()
+
+/**
+ * Extract the Amazon domain from cookies
+ * Returns the domain without the leading dot (e.g., "amazon.com", "amazon.co.uk", "amazon.de")
+ */
+export function getAmazonDomain(): string {
+  if (!AMAZON_COOKIES || AMAZON_COOKIES.length === 0) {
+    console.error('[WARN] No cookies found, using default amazon.com domain')
+    return 'amazon.com'
+  }
+
+  // Find a cookie with domain starting with ".amazon."
+  const amazonCookie = AMAZON_COOKIES.find(cookie => 
+    cookie.domain && cookie.domain.startsWith('.amazon.')
+  )
+
+  if (amazonCookie) {
+    // Remove the leading dot from domain
+    const domain = amazonCookie.domain.startsWith('.') 
+      ? amazonCookie.domain.substring(1) 
+      : amazonCookie.domain
+    console.error(`[INFO] Detected Amazon domain from cookies: ${domain}`)
+    return domain
+  }
+
+  // Fallback: try to find any cookie with "amazon" in the domain
+  const fallbackCookie = AMAZON_COOKIES.find(cookie => 
+    cookie.domain && cookie.domain.includes('amazon')
+  )
+
+  if (fallbackCookie) {
+    let domain = fallbackCookie.domain
+    // Remove leading dot if present
+    if (domain.startsWith('.')) {
+      domain = domain.substring(1)
+    }
+    // If it's a subdomain like "www.amazon.com", extract the main domain
+    if (domain.startsWith('www.')) {
+      domain = domain.substring(4)
+    }
+    console.error(`[INFO] Detected Amazon domain from cookies (fallback): ${domain}`)
+    return domain
+  }
+
+  console.error('[WARN] Could not detect Amazon domain from cookies, using default amazon.com')
+  return 'amazon.com'
+}
